requirements.txt
httpx>=0.25.0
colorlog>=6.7.0
vdf>=3.4.0
aiofiles>=23.2.1
fastapi>=0.109.0
uvicorn>=0.27.0
pywebview>=4.3.0
jinja2>=3.1.3
python-multipart>=0.0.6
pydantic>=2.5.0
